+++ src/app/modules/resource-table/resource-table.component.ts
@@ -23,6 +23,7 @@ import {
 import { ColumnDescription } from '../../types/column.description';
 import { distinctUntilChanged, sample, tap } from 'rxjs/operators';
 import {
+  dispatchWindowResize,
   escapeStringForRegExp,
   getPluralFormOfRecordName, getPluralFormOfResourceType
 } from '../../shared/utils';
@@ -42,7 +43,7 @@ import { FilterType } from '../../types/filter-type';
 import { CustomDialog } from '../../shared/custom-dialog/custom-dialog.service';
 import Resource = fhir.Resource;
 import { MatExpansionPanel } from '@angular/material/expansion';
-import { MatLegacyTooltip as MatTooltip } from '@angular/material/legacy-tooltip';
+import { MatTooltip } from '@angular/material/tooltip';
 import { CdkVirtualScrollViewport } from '@angular/cdk/scrolling';
 import { isEqual, pickBy } from 'lodash-es';
 import { saveAs } from 'file-saver';
@@ -641,6 +642,9 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
    */
   toggleFullscreen(): void {
     this.fullscreen = !this.fullscreen;
+    setTimeout(() => {
+      dispatchWindowResize();
+    }, 200);
   }
 
   /**
@@ -689,7 +693,7 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
     }
 
     // MatTable shows sort order icons in reverse (see comment to PR on LF-1905).
-    const isAsc = this.sort.direction === 'desc';
+    const isAsc = this.sort.direction === 'asc';
     const allColumns = this.columnDescriptionsService.getAvailableColumns(
       this.resourceTypeColumns || this.resourceType,
       this.context
@@ -921,8 +925,7 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
       message = `The data was sorted by ${
         sortingColumnDescription.displayName
       } in ${
-        // MatTable shows sort order icons in reverse (see comment to PR on LF-1905).
-        this.sort.direction === 'desc' ? 'ascending' : 'descending'
+        this.sort.direction === 'asc' ? 'ascending' : 'descending'
       } order.`;
     }