+++ src/app/modules/search-parameter/search-parameter.component.spec.ts
@@ -10,13 +10,12 @@ import {
 import { CommonModule } from '@angular/common';
 import { MatIconModule } from '@angular/material/icon';
 import { ReactiveFormsModule } from '@angular/forms';
-import { MatLegacyButtonModule as MatButtonModule } from '@angular/material/legacy-button';
-import { MatLegacyFormFieldModule as MatFormFieldModule } from '@angular/material/legacy-form-field';
-import { MatLegacyAutocompleteModule as MatAutocompleteModule } from '@angular/material/legacy-autocomplete';
-import { MatLegacyInputModule as MatInputModule } from '@angular/material/legacy-input';
+import { MatButtonModule } from '@angular/material/button';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatAutocompleteModule } from '@angular/material/autocomplete';
+import { MatInputModule } from '@angular/material/input';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { AutocompleteModule } from '../autocomplete/autocomplete.module';
-import { RouterTestingModule } from '@angular/router/testing';
 
 class Page {
   private fixture: ComponentFixture<SearchParameterComponent>;
@@ -51,8 +50,7 @@ describe('SearchParameterComponent', () => {
           MatAutocompleteModule,
           MatInputModule,
           NoopAnimationsModule,
-          AutocompleteModule,
-          RouterTestingModule
+          AutocompleteModule
         ]
       },
       {