+++ src/app/modules/step-0-settings-page/fhir-server-select/fhir-server-select.component.ts
@@ -13,7 +13,7 @@ import {
   Self,
   ViewChild
 } from '@angular/core';
-import { MatLegacyFormFieldControl as MatFormFieldControl } from '@angular/material/legacy-form-field';
+import { MatFormFieldControl } from '@angular/material/form-field';
 import { BaseControlValueAccessor } from '../../base-control-value-accessor';
 import { AbstractControl, NgControl } from '@angular/forms';
 import { Subject } from 'rxjs';