+++ src/app/shared/column-descriptions/column-descriptions.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatLegacyDialogModule as MatDialogModule } from '@angular/material/legacy-dialog';
+import { MatDialogModule } from '@angular/material/dialog';
 import { SelectColumnsModule } from '../../modules/select-columns/select-columns.module';
 
 @NgModule({